+++ source/be_decompress.h
@@ -1,4 +1,5 @@
 #ifndef BE_DECOMPRESS_H_
 #define BE_DECOMPRESS_H_
-object decompress(unsigned int c);
+#include <stdint.h>
+object decompress(uintptr_t c);
 #endif
